Move IGameServiceProvider into its own header file
