+++ src/game_logic/ai/prisoner.cpp
@@ -24,7 +24,7 @@
 #include "game_logic/damage_components.hpp"
 #include "game_logic/entity_factory.hpp"
 
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 
 namespace rigel { namespace game_logic { namespace ai {
