+++ src/game_logic/ai/sliding_door.cpp
@@ -22,7 +22,7 @@
 #include "engine/base_components.hpp"
 #include "engine/entity_tools.hpp"
 #include "engine/physical_components.hpp"
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 
 namespace rigel { namespace game_logic { namespace ai {
