+++ src/game_logic/ai/slime_pipe.cpp
@@ -21,7 +21,7 @@
 #include "engine/physical_components.hpp"
 #include "game_logic/entity_factory.hpp"
 #include "game_logic/damage_components.hpp"
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 
 namespace rigel { namespace game_logic { namespace ai {
