+++ src/game_logic/interaction/elevator.cpp
@@ -19,7 +19,7 @@
 #include "engine/entity_tools.hpp"
 #include "engine/physical_components.hpp"
 #include "game_logic/player/components.hpp"
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 namespace ex = entityx;
