+++ src/game_logic/item_container.cpp
@@ -26,7 +26,7 @@
 #include "game_logic/effect_components.hpp"
 #include "game_logic/entity_factory.hpp"
 
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 
 namespace rigel { namespace game_logic {
