+++ src/game_logic/player/attack_system.ipp
@@ -16,7 +16,7 @@
 
 #include "data/player_data.hpp"
 #include "game_logic/player/attack_traits.hpp"
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 #include <cassert>
