+++ src/game_logic/player/projectile_system.cpp
@@ -23,7 +23,7 @@
 #include "game_logic/damage_components.hpp"
 #include "game_logic/entity_factory.hpp"
 
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 
 namespace rigel { namespace game_logic { namespace player {
