+++ src/game_logic/player_interaction_system.cpp
@@ -24,7 +24,7 @@
 #include "game_logic/damage_components.hpp"
 #include "game_logic/interaction/force_field.hpp"
 #include "game_logic/interaction/teleporter.hpp"
-#include "game_mode.hpp"
+#include "game_service_provider.hpp"
 
 
 namespace rigel { namespace game_logic {
