+++ src/game_session_mode.cpp
@@ -18,6 +18,8 @@
 
 #include "base/warnings.hpp"
 
+#include "game_service_provider.hpp"
+
 RIGEL_DISABLE_WARNINGS
 #include <atria/variant/match_boost.hpp>
 RIGEL_RESTORE_WARNINGS
