+++ src/intro_demo_loop_mode.cpp
@@ -2,6 +2,8 @@
 
 #include "loader/resource_loader.hpp"
 
+#include "game_service_provider.hpp"
+
 
 namespace rigel {
