+++ src/ui/apogee_logo.cpp
@@ -18,6 +18,8 @@
 
 #include "loader/resource_loader.hpp"
 
+#include "game_service_provider.hpp"
+
 
 namespace rigel { namespace ui {
