+++ src/ui/bonus_screen.cpp
@@ -20,6 +20,8 @@
 #include "loader/resource_loader.hpp"
 #include "ui/utils.hpp"
 
+#include "game_service_provider.hpp"
+
 
 namespace rigel { namespace ui {
